+++ packages/core/src/index.ts
@@ -34,7 +34,6 @@ export * from "./import"
 export * from "./fs"
 export * from "./tokens"
 export * from "./promptdom"
-export * from "./filedom"
 export * from "./search"
 export * from "./yaml"
 export * from "./mime"
