+++ packages/core/src/promptcontext.ts
@@ -10,6 +10,7 @@ import { readText } from "./fs"
 import {
     PromptNode,
     appendChild,
+    createDefDataNode,
     createFileMergeNode,
     createFunctionNode,
     createImageNode,
@@ -17,7 +18,6 @@ import {
     createSchemaNode,
 } from "./promptdom"
 import { bingSearch } from "./search"
-import { createDefDataNode } from "./filedom"
 import { CancellationToken } from "./cancellation"
 import {
     RunPromptContextNode,
@@ -64,7 +64,7 @@ export function createPromptContext(
         stringify: INIStringify,
     })
     const XML = Object.freeze<XML>({
-        parse: XMLParse
+        parse: XMLParse,
     })
     const AICI = Object.freeze<AICI>({
         gen: (options: AICIGenOptions) => {
@@ -164,8 +164,8 @@ export function createPromptContext(
 
     const ctx = Object.freeze<PromptContext & RunPromptContextNode>({
         ...createRunPromptContext(options, env, trace),
-        script: () => { },
-        system: () => { },
+        script: () => {},
+        system: () => {},
         env,
         path,
         fs,
@@ -191,7 +191,7 @@ export function createPromptContext(
             throw new CancelError(reason || "user cancelled")
         },
         defData: (name, data, defOptions) => {
-            appendPromptChild(createDefDataNode(name, data, env, defOptions))
+            appendPromptChild(createDefDataNode(name, data, defOptions))
             return name
         },
         fetchText: async (urlOrFile, fetchOptions) => {
